Addition of DVC listing search tool
This script allows the user to automatically look for dvc listings
and receive an email for new listing which match the user's filter.

The script is pretty simple; scrap the current dvc listings from
www.disneydvcresale.com, and find any IDs which we haven't yet
seen. The disney dvc resale site already aggregates results from
OTHER resale sites, so it's rather nifty. There does not appear
to be a public API, so I just resorted to scraping the raw page
source for my purposes.

You'll also need to either set up an snmp server to shoot off
an email, or (what I did) sign up with mailgun (which is free
for a certain number of emails per month). They will give you
an API key which you can plug into the structure I already
have.

This is meant to be used alongside crontab, which is why it's a
simple one-shot program. I didn't bother to add any CLI params
either, but that's just because I made this for personal use.